+++ example/lib/annotation_order_maps.dart
@@ -51,12 +51,12 @@ class _AnnotationOrderBodyState extends State<AnnotationOrderBody> {
                         target: center,
                         zoom: 5.0,
                       ),
-                      annotationOrder: const [
+                      annotationOrder: const {
                         AnnotationType.line,
                         AnnotationType.symbol,
                         AnnotationType.circle,
                         AnnotationType.fill,
-                      ],
+                      },
                     ),
                   ),
                 ),
@@ -85,12 +85,12 @@ class _AnnotationOrderBodyState extends State<AnnotationOrderBody> {
                         target: center,
                         zoom: 5.0,
                       ),
-                      annotationOrder: const [
+                      annotationOrder: const {
                         AnnotationType.fill,
                         AnnotationType.line,
                         AnnotationType.symbol,
                         AnnotationType.circle,
-                      ],
+                      },
                     ),
                   ),
                 ),