+++ example/lib/layer.dart
@@ -54,7 +54,7 @@ class LayerState extends State {
                 target: center,
                 zoom: 11.0,
               ),
-              annotationOrder: const [],
+              annotationOrder: const {},
             )),
       ),
       TextButton(