+++ maplibre_gl_platform_interface/lib/src/ui.dart
@@ -83,7 +83,10 @@ class CameraTargetBounds {
 @immutable
 class MinMaxZoomPreference {
   const MinMaxZoomPreference(this.minZoom, this.maxZoom)
-      : assert(minZoom == null || maxZoom == null || minZoom <= maxZoom);
+      : assert(
+          minZoom == null || maxZoom == null || minZoom <= maxZoom,
+          'minZoom cannot be larger than maxZoom.',
+        );
 
   /// The preferred minimum zoom level or null, if unbounded from below.
   final double? minZoom;