+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * Renamed `Maplibre` to `MapLibre` to be in line with maplibre-native 
   (affects for example the classes MaplibreMap and MaplibreMapController).
 * Remove the parameter `hash` from `MapOptions` because it was not used.
+* The parameter type of  `annotationOrder` and `annotationConsumeTapEvents` from 
+  `MapLibreMap` has changed to a Set.
 
 ### Changes